move DeployMetadata to utils.py
